requirements.txt
pymupdf
groq
python-dotenv
pymupdf4llm
llama-index
textgrad